Fix flash of unstyled content when switching from Auto to a specific theme (#44)
* Lint with prettier

* Remove reference to specific theme ID, replacing with "enum" value

* Fix flash on unstyled content when changing from auto to specific theme

Most browsers don't download CSS files with `media` attributes that don't apply to the current rendering context. Because of this, when the original link tag is removed, there is no CSS styling for the page for a split second (or multiple seconds on a slower connection).

Instead of immediately removing the old `link` tag, we now add a brand new link to the desired CSS, along with an `onload` event handler. This way we only remove the old `link`s after the new theme CSS has been loaded.

Using `onload` on the existing (but not used) `link` doesn't seem to work (at least in Chrome) as the even never gets triggered.